Добавил экран авторизации и соответствующий флоу.

+++ data/navigator/src/main/java/com/lofigroup/data/navigator/local/ProfileDataSourceImpl.kt
@@ -29,7 +29,7 @@ class ProfileDataSourceImpl @Inject constructor(
   }
 
   private fun generateNewProfile(): ProfileDataModel {
-    val id = Random.nextInt()
+    val id = Random.nextLong()
 
     val profile = ProfileDataModel(id = id, name = "No name")
     val json = adapter.toJson(profile)
